Move disposable eltoo fixtures out of the contract module
Move the existing fixture body into its own file without changing its
keys, funding, or generated terms. This separates the fixture move from
the following contract and runner refactor.

Validation: integration library and all eight eltoo security tests pass.
